Exclude i18next type customization from built package (#7467)
We have a module augmentation for the i18next module to make it so
that calls to `useTranslation` and friends are typechecked to make
sure they are correct.  But we don't want to ship that module
augmentation, because that means that users of our package can't
specialize it for their own use.  So instead move it into a .d.ts
file, which will not make its way into the rollup build.

Test Plan:

npm run build in opensource/sdk-js, saw that there was no definition
of CustomTypeOptions in the output files for packages/types.

Changed a call to translations to a key that doesn't exist, it still
caused a compile error.

monorepo-commit: 75ec3d60aa26489a195546d2ec15b698002e933e