ThreadedComments: Adding a different highlight color to resolved threads
If a resolved thread in ThreadedComments was highlighted, we would correctly apply the `highlighted` classname to the ThreadedComments thread, however it would not appear any different to the user, as all resolved threads were already greyed out with the color used to highlight open threads (which have a white background normally).

This PR makes sure users can distinguish if a resolved thread is highlighted, by making it a shade of grey darker than the rest of the resolved threads.

Note: The specificity of the selector applied is still (0, 1, 0), so it will not break any existing implementations.

Test Plan:
The image below shows 3 resolved threads, out of which the middle one is
highlighted.

![image](https://github.com/getcord/monorepo/assets/102963329/83e10bca-5f55-452e-bb46-b660348f354a)



Reviewers: AlbertQM

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/7474

monorepo-commit: b458f11c12996f3b4f7f5671e8ec289b776b11bf