Experimental Facepile
Removed the `showPresence` prop altogether, as `Facepile` never uses it. It's `PresenceFacepile` that has presence information! This simplified the code and removed the need to convert `AvatarWithPresence`.

Test Plan:
Use the new Facepile in the docs, it renders the same thing as staging/prod


Reviewers: lazybean, Netceer

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/7468

monorepo-commit: 1c6cfd9ddb2be34806be58050eaec3edf8ce3cd7