Filter messages by authorID (#7480)
## Summary

The GET messages REST API endpoint has filters but they support very
limited options. I'm interested on getting the messages created by a
given user, and it's timing out for me when fetching a big amount of
messages.

In addition to debugging the issue with the timeout, which I'll do, I
think it'd make sense to add a filter for authorID so that we don't need
to fetch so many messages in the first place and then filter out.

monorepo-commit: eb2c1f0172a6b680e3d2edbce69d4d56156d467a