withWrapper -> withCordClassname (#7524)
summary:
We were adding an extra div with .cord-component and
.cord-component-<name>
The reason initially was because new nonion component should be
replacing the old one (updating inline).
But css would break, because of the DOM structure change.

Turns out that:
- DOM structure would change and css would break (we add extra div to
  every sub component as well)

- we do not update old code but create new component, so breaking
  selectors is OK.

  This div removes the extra div (wrapper) and only apply
  `.cord-component` class.

Test Plan:
See test bed, extra div is gone, some css is borked

Reviewers:

monorepo-commit: 2d17316f0970c23f6e5bc19baf7bbc767cf3070f