Add base code for translating notifications (#7521)
This adds the API parameters and plumbing through the GraphQL layer
for translating notifications.  It currently just sends null all the
time for the translation data, that will come along in a later diff.

Since notification bodies are generated at runtime anyway, we don't
need to change the DB, we just need to send down the translation key
and any parameters that are needed.

Senders are automatically included in the translation parameters
during the API conversion process, so we don't need to send those down
via GraphQL.  Instead we send down SimpleTranslationParameters, which
is all the parameters except the users involved.

Test Plan:

Fetch https://local.cord.com:8161/v1/users/flooey/notifications and
run cord.notification.observeData() on the testbed.  Both included a
null translation key and parameters.

monorepo-commit: 115a0c768f83434a6a339d6183df79ab4e0b123e