Add context to i18n strings (#7556)
Add more context to each string in the i18n list so that it's easier
for translators to provide correct translations.

Also move the "show more" strings from the `message` namespace to the
`thread` namespace, as message pagination is really part of the thread
behavior (these strings cannot appear when using a `cord-message`
component, for instance).

Test Plan:

Mark I Eyeball

monorepo-commit: 5292bd433af0dfc9f45f964ce739a1c59966f349