Adding a filter to useThread
When highlighting a thread in ThreadedComments (#7457), we fetch this thread separately, using the `useThread` hook, just in case it's paginated. We need to make sure that we don't just display any threadId that we get as a prop. The thread we fetch has to match the filters of the ThreadedComments (location, partialMatch, groupID and metadata). Being able to just have the useThread hook apply those filters for us, instead of us checking each one, is a useful thing to include in the API, as a customer may run into a similar use case.

This PR adds a filter parameter to the options of the `useThread` hook. It re-uses the `ClientThreadFilter` type of the `useThreads` hook, for consistency.

The filtering is done on the client side, because we want to take advantage of fetching the thread from the context, if it's already there.

Test Plan:
Used a thread with the following attributes:
* location: { page: 'testbed' }
* groupID: 'cord'
* resolved: false
* metadata: { abc: '123' }

In the browser console, fetched this thread with the following filters:

* `filter: { location: {} }` - returned null
* `filter: { location: { page: 'testbed' } }` - returned the thread
* `filter: { location: { value: {} } }` - returned null
* `filter: { location: { value: {}, partialMatch: true } }` - returned the thread
* `filter: { location: { value: {}, partialMatch: false } }` - returned null
* `filter: { location: { value: {}, partialMatch: 'true' } }` - returned null
* `filter: { groupID: 'cord' }` - returned the thread
* `filter: { groupID: '' }` - returned the thread (considered undefined)
* `filter: { groupID: 'cord2' }` - returned null
* `filter: { metadata: {} }` - returned the thread
* `filter: { metadata: { abc: '123' } }` - returned the thread
* `filter: { metadata: { random: 'metadata' } }` - returned null
* `filter: { resolvedStatus: 'any' }` - returned the thread
* `filter: { resolvedStatus: 'unresolved' }` - returned the thread
* `filter: { resolvedStatus: 'resolved' }` - returned null


Reviewers: dmmiller, flooey, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/7538

monorepo-commit: 86f63b19ea41a22a7978f72b98e179a213612ef8