Experimental PresenceObserver
Copy pasted our current PresenceObserver, and made the adjustments needed for it to work outside the onion.

Test Plan:
use `experimental.PresenceObserver` in our internal PagePresence, in the PresenceFacepile docs and in the testbed.
Presence seems to work (both in PagePresence where we observe the document visibility, and in other places where we just observe presence in a specific element).


Reviewers: Netceer, lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/7564

monorepo-commit: ad5fa4a2c2e6725b3e602b777616d55c1dbf5e7f