ThreadedComments: Showing highlighted thread even if paginated
Currently in ThreadedComments, if a developer attempts to highlight a threadID that is paginated, none of the threads rendered will be highlighted. It will only show the highlighted thread when a user manually fetches more threads and reaches the highlighted one. We should always attempt to fetch and show the highlighted thread, unless it doesn't exist.

This PR always fetches the highlighted thread, and scrolls it into view. In the case when the thread is paginated, we display the thread at the end of the ThreadedComments thread list, until the user fetches enough threads for it to be in its correct sort position. We also take care to highlight resolved threads and set the initial tab to resolved, if necessary.

I have made a small tweak in the props of the ThreadedCommentsThreadList component (not exposed to customers). Specifically, I am passing a full ThreadSummary, instead of a `highlightThreadId` string, so we don't need to fetch the highlighted thread again.

I have also created a local state of threads within the ThreadedCommentsThreadList. When adding a highlighted thread into our thread list shouldn't mutate the hook result, but rather keep a separate state of threads to render. I have made that change in the ThreadedCommentsThreadList.

Test Plan:
Tested all permutations of the following:

`displayResolved`
* interleaved
* sequentially
* tabbed
* resolvedOnly
* unresolvedOnly

`highlighted thread`
* unresolved
* unresolved and paginated
* resolved
* resolved and paginated

In all cases, the highlighted thread appeared scrolled into view.
When the highlighted thread was resolved in the tabbed view, the initial tab was the resolved tab.
When the highlighted thread was resolved In the sequential view, the resolved section was initially expanded.


Reviewers: flooey, lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/7457

monorepo-commit: f13f5022ed153d314ec3be40deb16b00422a5a03