Move `dayjs` import outside the opensource package
The current opensoruce package is broken because of `dayjs` not being correctly imported in it.
The CI/npm package build somehow didn't fail and we ended up with a version (1.28) with crashes on load.
Reverting all of this while we try to figure more stuff out.

Test Plan:
Load an experimental component, check it shows "dummy info".
Load a PagePresence, hover on the avatars, it works.


Reviewers: lazybean, flooey

Reviewed By: lazybean, flooey

Pull Request: https://github.com/getcord/monorepo/pull/7592

monorepo-commit: 707eaa5780cff3e0ea1e39293073280183f08179