Remove ts-ignore declarations (#7606)
We have a bunch of ts-ignore declarations in the opensource code
because we're importing React in order to have it available to the JSX
transform, but not using it as a value.  That works fine for our
opensource code, but the main server tsconfig includes
`"importsNotUsedAsValues": "error"`, which means that it is an error
to do this.

However, we don't really need to have that flag.  We mostly have it
because we wanted to alert ourselves to places where we were importing
a type and wanted to be clear about it, but we also use the
`@typescript-eslint/consistent-type-imports` line rule which serves
the same purpose.  So switch `importsNotUsedAsValues` to `preserve`,
which maintains the same behavior but doesn't cause an error for an
unused import and remove the `@ts-ignore` declarations.

Test Plan:

Loaded the testbed, ThreadedComments, Thread, and Sidebar all rendered
properly.

Built sdk.latest.js under both configurations and compared them.  They
produce the same code, modulo things like timestamps.

monorepo-commit: ea4912f5c1a28e56de136d0b86731cd77ef40ca4