FileAttachment: move Remove/Download button CSS to own files
Fixing two long standing TODOs, because I need to de-onion this component.

When you upload a file, we show a FileAttachment. If you hover on it, the "Remove" button appears.
When you send a message with an attachment, we show a MessageFileAttachment. If you hover on it, the "Download" button appears.

removed the `internalStyle`, as it wasn't used anymore.

Opensourced FileAttachment.classnames, as I'll need to opensource components importing those classnames soon.

Test Plan:
Upload a file, hover on it.
Send the message, hover on the attachment.
Everything behaves as in prod.


Reviewers: lazybean, Netceer

Reviewed By: Netceer

Pull Request: https://github.com/getcord/monorepo/pull/7633

monorepo-commit: 603f08deabace84dfce8418590175d54c0f923cf