Experimental EmojiPicker and WithPopper
Both are needed for Reactions so moving them to opensource.

Mainly copy paste with a couple of more deliberate changes.

1. Removed useWidthTracker on WithPopper that relied on ResizeObserverContext
We were using it to find the width of the element the popper was coming
out of it and to then set the popper width to match it (composer uses
this to have the @mention list take same width).
Floating UI library has a size middleware that can access the attached
element and use the width to for the popper element itself.

2. removed setting up embedXFrame listeners in useClickOutside
This might need to be added back in by copy pasting all the embedXFrame
code and types but it mainly looked like it was being used for making
annotations work across iframes?


Reviewers: AlbertQM

Reviewed By: AlbertQM

Pull Request: https://github.com/getcord/monorepo/pull/7600

monorepo-commit: b8af2ca886b0e2f26dbcc627801b8b5b245a2eba