Add `addSubscribers` and `removeSubscribers` to the Update thread REST API
In https://github.com/getcord/monorepo/pull/7191, we added the ability to specify subscribers at thread creation.  This adds the ability to add and remove subscribers in thread update.

This is only supported on the REST API since it doesn't make sense to allow people to add/remove other subscribers from the user side and they can already subscribe or unsubscribe themself via JS API.


Test Plan:
Created a thread
Using REST endpoint, added and removed subscribers to it, verified in the db that the correct value was set

I used the same user in add and remove, got an error
I used a non existent user and it was ignored [This is the same as other places like typing]

Reviewers: flooey, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/7685

monorepo-commit: 4bb16c22674a7c1eeea24ac903f3081859e3807e