MessageOptions: Never hide if open
The MessageOptions menu buttons appear on hover (on `message:hover`).
If you stop hovering on the message, the MessageOptions disappear.
This is usually hard to do (because we have an overlay as big as the page, so you can't easily stop the `:hover`). However, in some scenarios, this is not true. E.g. if Cord is rendered inside an iframe, the overlay will just be as big as the iframe -- now it's much easier to stop `:hover`ing.
This was spotted by datapeople in our Docs's "Ask Cordy" chat.

Test Plan:
Before, the menu is hidden if we stop hovering, even when it was open!

After, if the menu is open, we never hide it.


Reviewers: dmmiller, lazybean

Reviewed By: dmmiller

Pull Request: https://github.com/getcord/monorepo/pull/7776

monorepo-commit: c32ae04000a151a3d54b975dd0140a57a5065256