Allow bullets to be indented (#7782)
Allow bullets, both normal and numbered, to be indented up to 4 times
(5 total levels of indentation).

The max indentation here is mostly a practical matter in terms of how
much different CSS we want to generate and dealing with the bullets
running off the right edge of the composer if it's too narrow for the
indentation.  5 levels of indentation seemed like a nice balance.

The data model change here is very small: only a new optional `indent`
property on bullets and numbered bullets.  If not present, the default
value is 0, meaning no indentation.  This means indented bullets are
*not* nested bullets, structurally, they are bullets that are
indented.  This displays the same to the user, but it makes the data
model simpler and easier to manage (always just a list of bullets).

The bullets end up producing the technically-invalid HTML sequence
`<ul><ul><ul><li>Example</li></ul></ul></ul>` (there should be `li`
elements interleaved with the `ul` elements), but this is also the
format produced by copying-and-pasting Google Docs indented bullets
and is well supported in all browsers I tested (Chrome, Firefox,
Safari).

Test Plan:

<img width="309" alt="Screenshot 2024-01-09 at 4 24 31 PM" src="https://github.com/getcord/monorepo/assets/215991/02649344-0492-4ad1-a821-a564bd6f0942">

<img width="301" alt="Screenshot 2024-01-09 at 4 47 14 PM" src="https://github.com/getcord/monorepo/assets/215991/e9ed6847-9fea-4de1-8769-ff051bc88e56">

monorepo-commit: 421f3fd8c887935a5a8f1b07a848218ef9d32d9e