Opensource MessageContent
This is the de-onion version of `cord-message-content`.
I was able to remove all internal contexts as they don't apply to the two de-onion components (MessageText and MessageFilesAttachments).

Test Plan:
Use it both in testbed (so I can try it out with image attachments), and in the docs. It looks OK.


Reviewers: lazybean, Netceer

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/7865

monorepo-commit: bc9b4e375f001cd9033b5255556df650b4e0adef