Rewrite imports to full paths
This is a 100% mechanical change to resolve imports to full paths.
While esbuild is perfectly happy with these extension-less paths (which
sometimes omit `index.ts`) in all modes, Node is not -- and although our
builds will continue to rely on esbuild, we do want to turn on the most
restrictive tsconfig settings for maximum compatibility (and to catch
issues with *external* imports) -- so we need to do this.


Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/7894

monorepo-commit: 6a03ab58fc467b20a6aadb6e1fb566abfa924c0e