Remove ANNOTATION message nodes (#7951)
We haven't allowed `ANNOTATION`-type nodes to be created in messages
for approximately 9 months, instead annotations are displayed below
the message as an attachment and don't appear in the message content.

In order to simplify things in the future, this removes `ANNOTATION`
from the possible message contents.  To handle existing old messages
that have these nodes, they are stripped from existing messages.  This
is done at runtime by cleansing messages as they get loaded from the
database.  This can serve as a model for lots of migrations we might
want to do in the future.

If we wanted, we could run an async job to gradually load and resave
any affected messages so that we could eventually remove this from the
system permanently, but it's not strictly necessary.

Test Plan:

Loaded a thread with an in-message annotation in a `cord-thread` in
the testbed.  The annotation displayed properly below the message.

Left FloatingComments messages, they displayed properly.

Send a message from the sidebar with an annotation, it displayed
properly and the arrow showed up correctly.

monorepo-commit: caa94a3b27680f6bf0bfd61abd0017139d9349ea