Add annotation placement to MessageAnnotationAttachment (#7940)
Add information about annotation placement to
MessageAnnotationAttachment.  This should make it easier to migrate
from FloatingThreads to other implementations or enable customer code
to place pins.

I think the names might be able to be improved, definitely interested
in thoughts on that front.


Test Plan:

Created a non-custom annotation in the testbed, then fetched it with
the JS and REST APIs, they both came through.

Created a custom annotation in the testbed, then fetched it with the
JS and REST APIs, they both came through.

monorepo-commit: 493cce1f7c2841c0457d64d6b92295ba24e9f328