add onClose to Menu (#7982)
Summary:
If we want user to be able to add item to a menu they need a way to
close it.

We do not do it by default(event delegation on the container), so we can support submenu.
Alernative would be to have user stopping event propagation.

Test Plan:
added a item with submenu, could close the whole menu

Reviewers:

monorepo-commit: e8c1bd030652e4b6ac74a2f6c0477de6d4ff17dc