ThreadedComments: Refactoring the highlighted thread logic
When highlighting a thread in ThreadedComments we check whether the ThreadedCommentsThreadList is rendering open or resolved threads, to then understand whether we should render the highlightedThread. This is a chunky switch/case statement which introduces a new variable and isn't the cleanest implementation.

We can simplify the code quite a lot using the filter parameter of the `useThread` hook (added in #7538). This PR refactors the thread highlighting implementation to make sure that individual threads always match the filter of the thread list they are rendered in.

Test Plan:
Highlighted the following threads in ThreadedComments:
1. An open and paginated thread (✅ Was highlighted correctly, the thread did not appear when switching to the resolved tab)
2. A resolved and paginated thread (✅ Was highlighted correctly and ThreadedComments showed the resolved tab from in the initial state - when switching to the open tab, the highlighted thread was not rendered)
3. An open thread in a partially matched location
    ✅  When ThreadedComments had `partialMatch={false}`, the thread did not appear
    ✅  When ThreadedComments had `partialMatch={true}`, the thread appeared


Reviewers: dmmiller

Reviewed By: dmmiller

Pull Request: https://github.com/getcord/monorepo/pull/8022

monorepo-commit: 7cbc6f2484ff060458c06c9d7f4f5fbf8e414250