ThreadedComments: Displaying highlighted threads
At the moment, when we highlight a paginated thread in ThreadedComments, it will appear within the ThreadedComments thread list. However, if the `highlightThreadId` prop changes, the thread disappears.

This PR makes sure that ThreadedComments displays all threads that have been highlighted within each render cycle. We do that by keeping a state of highlighted threads. So when the prop changes, we still have all threads that have been highlighted in the current render-cycle.

In the case where highlighted threads are paginated, we just append them to the end of the current thread list. However, since we now can have multiple threads which have been highlighted within the render-cycle, we need to make sure that the appended threads are in the correct order. So I opted to sort them in the client-side every time we add a new highlighted thread in the state.

Test Plan:
The attached video does the following:

1. Highlights the thread by andrei (July 1, 2022) for 3 seconds (originally is paginated, so it is brought into view)
2. Removes the highlight and keeps the `highlightThreadId` prop empty for 3 seconds
3. Highlights the thread by flooey (Jan 18, 2023) for 3 seconds (also paginated, but dated after the first highlighted thread)
4. Loads all threads until the highlighted ones reach their original position
5. Clicks on the `toggle metadata` button. This button applies `filter: { metadata: { page: 'one' } }` to ThreadedComments (Only 1 thread renders)
6. Switches between open and resolved tabs
7. Clicks on the `toggle groupdId` button, which changes the groupID of ThreadedComments to 'cord-one', which completely stops rendering the component.

https://github.com/getcord/monorepo/assets/102963329/8e3c5e3c-e2e3-490a-95bc-885c14fe5add



Reviewers: lazybean, flooey

Reviewed By: lazybean, flooey

Pull Request: https://github.com/getcord/monorepo/pull/7867

monorepo-commit: 86da5202cb1289409974af3dc7421aa9a1992877