Clean up list group members and add parameter type
We had a customer who had an issue with fetching group members and
it turns out they had 1001 members in the group, and the results
were only showing 1000.

This PR does the following:
- Relocates the new type from organization.ts to group.ts (my bad)
- Adds a new type for parameters as I realised this was missing.
- Making sure we state in the `limit` parameter it defaults to 1000.

Will also update all other endpoints with this default, and eventually
the cli (the description that is)

Test Plan:
Can see it in the docs locally!


Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/8109

monorepo-commit: 0d78f0eb8a518475ee370d017c881f1931f67ba6