Add setSeen(filter) API call (#8054)
Adds a new API call that lets the caller mark an entire collection of
threads as seen or unseen.  This is a new overload of `setSeen()` that
takes a filter object, and it marks every thread that matches the
filter as seen or unseen.  The existing overload that takes a single
threadID is still in place, so existing code will continue to work.



Test Plan:

https://github.com/getcord/monorepo/assets/215991/aefdbe52-32ea-467c-9d5e-8677a7372176

monorepo-commit: 8d4b6509c4c4eb9b2811679d3da7e1fefa2ddd70