Fix React types build (#8216)
In a few places, we were importing from a file inside the
`@cord-sdk/types` package, which then the build couldn't recognize as
an external import and tried to internalize, which ended up meaning
the types ended up in the wrong place.  Fix all those and add a lint
rule that prevents it.

Test Plan:

`npm run build` puts the types in the right place in the
`@cord-sdk/react` build.

monorepo-commit: 3cf25a51a99bcaeafb9ccd7596540226b9999643