Expand reaction tooltip messages (#8355)
In some languages, the verb conjugation that you need to use in the
reaction tooltip differs depending on whether the viewer is included
and how many users reacted, so expand the set of messages to include
pluralization options and to differentiate between user sets that
include the viewer and ones that don't.


Test Plan:

Looked at various combinations of reactions in a Reactions component,
the tooltips look the same as they did before this diff.

monorepo-commit: 02ef3b8a379288dacbe07e9700eb87b6501a2bf5