Allow ThreadedComments to not have groupID if composerPosition is 'none'
ThreadedComments is a component where it would be very useful to load
threads from multiple groups rather than tying it to one group ID.  The
reason that, until now, it has required a groupID, is so we know where
to write new threads created via the composer at the bottom.  If the
composer isn't being shown, we don't have to worry about this.


Reviewers: dmmiller

Reviewed By: dmmiller

Pull Request: https://github.com/getcord/monorepo/pull/8406

monorepo-commit: cdf881098375f9990e8ed2dc28f0cef99caa58f7