Add ability to filter by subscription status (#8439)
Add the ability to filter a group of threads to only threads the
viewer is subscribed to.

This requires a lot of changes that are mostly plumbing, because this
is the first thread filter that depends on the viewer's state, not
just the properties of the thread object, so all the places where the
filter is applied need to have access to the viewer as well.  The
actual action is in `ThreadLoader.ts`, `ThreadParticipantMutator.ts`,
and `convertToExternal/thread.ts`.

The `viewer` filter is an array that only accepts the single value
`subscribed` for now, but we imagine including different other options
as well (eg, `mentioned` or `unread`), hence the array.  Customers can
just supply a single value not inside an array for ease of use, though
(eg, `viewer: "subscribed"`)


Test Plan:

Used `cord.thread.observeThreads` with a collection of subscribed and
unsubscribed threads, they returned the correct values.

Used `cord.thread.observeThreadCounts` with a collection of subscribed
and unsubscribed threads, they returned the correct values.

Used `cord.thread.setSeen` to mark all subscribed threads as seen, it
worked properly and didn't mark unsubscribed threads as seen.

Used `cord.thread.observeThreads` with a subscribed filter and then
mentioned a user (which subscribes them to the thread) in a new thread
by a different user, the return value updated with the new thread.
Did it without mentioning them (so the new thread isn't subscribed),
no update.

Used `cord.thread.observeThreadCounts` and did the same as above.

monorepo-commit: db8bfb7b1ed41b9b80be7b61383593d6843a0af6