Composer 4 attachments: update file based on id rather than name
How uploading files work:
- Initialise upload async via public API
- At some point in the future, we'll get the URL pointing to that file. This means we then have to get a specific file and update it.

Before this change, to find the file that needed to be updated, we were relying on the name.
After this change, we're relying on the `id`.

This is generally a much better pattern, and fixes the issue where you can't upload the same image multiple times.

Test Plan:
- Upload a file via the UI
- Upload it again, it's a separate file (instead of overriding the existing)
- Paste an image 4 times in the composer, I get 4 separate attachments.
- Click on the X to delete an image deletes the right one.
- Send the message, attachment is correctly shown in the message


Reviewers: elenaperezrioja, lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/8411

monorepo-commit: c63f39352383fb0fa63e146671d45958edf58d47