Add filtering by viewer mentioned (#8463)
Add the ability to filter by whether the viewer is mentioned in the JS
API.  This expands the `viewer` filter to allow the value
`"mentioned"`.  If you specify multiple values in the `viewer` array,
they combined as an `OR`, so you can do "subscribed or mentioned", but
you can't do "subscribed and mentioned".

Test Plan:

Used `cord.thread.observeThreads` with a collection of subscribed,
mentioned, both, and neither threads, they returned the correct
values.

Used `cord.thread.observeThreadCounts` with the same threads, they
returned the correct values.

Used `cord.thread.setSeen` to mark all mentioned threads as seen, it
worked properly and didn't mark unmentioned threads as seen.

Used `cord.thread.observeThreads` with a mentioned filter and then
mentioned a user in a new thread by a different user, the return value
updated with the new thread.  Did it without mentioning them, no
update.

Used `cord.thread.observeThreadCounts` and did the same as above.

monorepo-commit: d2e6e59c40c0e7ba2ce6a0e34e1035e8cbc93d2b