SDK: MessageContent.edited should be optional
It says in our docs that `edited` is an optional boolean, and it also makes sense that it would be. Though when actually using `MessageContent`, it is a required prop. Here I'm updating the component prop type to make it optional.

I believe this is all I need to do? (and that we want to do this 👀 ) I ran docs-codegen and it didn't change anything (because the docs for this are written out instead of using generated types.


Test Plan:
Should be able to use MessageContent without passing the `edited` prop

Reviewers: flooey, AlbertQM, Netceer, lazybean, elenaperezrioja

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/8461

monorepo-commit: 9af3070087b4728e4c9e5b54352a5a9c8bf001db