Add bootstrap i18n to CordContext (#8467)
In order to support SSR, make it so that the CordContext can supply
i18next instances even without getting them from the SDK.

Add a separate `i18n` property to the `CordContext`.  This will be
`sdk.i18n` when the SDK object exists, but if not, we'll create a
`bootstrapI18n` to be used in the meantime.  That will be cleared when
the SDK does get inited for the first time.

Move a bunch of translation init functions into the `@cord-sdk/react`
package, so that the CordContext can use them to init the bootstrap
i18next instance.

Also because we dynamically import all the dayjs locale files, we need
to add `@rollup/plugin-dynamic-import-vars` to our build.

Test Plan:

Used `cord.updateOptions()` to update translations, they got rendered
in both `ThreadedComments` (a no-onion component) and `Thread` (an
onion component).

Manually inspected the build `@cord-sdk/react` index.js file, the
dynamic import is resolved to a structure that finds all the locale
files.

monorepo-commit: 9b83fe3d9cf4b36e94c081d12f32fac06cfb0cd8