Add bundle-CSS script to package opensource's CSS
### Why

For Cord 4.0 world, we want to version our CSS.
In Cord 4.0, we are opensourcing the entirety of the code of our components. Shipping CSS every day will make it awkward to maintain backwards compatibility.
Having CSS matching the JSX of the components will make it easier for everyone to update/downgrade as needed and get components that work and look as they are meant to.


Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/8310

monorepo-commit: 822f8295529831860b25e36d40dea0bbca7bf7e6