Don't use initReactI18next (#8561)
I looked into the code, and all `initReactI18next` does is set the
instance you used it on as the global `i18next` instance for the React
components and set its options as the default options for the React
components.  We explicitly don't want to do that, we want to leave the
default instance for use by our customers and have set our code up to
always pass the Cord i18next instance into the components.  So stop
using initReactI18next.

Test Plan:

Built the package and used it inside an app with its own translations.
Before this diff, the app translations got overwritten in some
circumstances by the Cord translations and messed everything up, after
this diff, it doesn't.

monorepo-commit: e76e2486e60a3638dc053f025e0a12298a3df1b7