Refactor avatar image status to support SSR
We're using useSyncExternalStore to support SSR within our avatar component. Currently we're being very optimisitic and asuming that if it's being server side rendered that it is loaded successfully.

On the client, we're checking the naturalWidth and complete properties of the image to check the status of the image.


Test Plan:
Ran it in a SSR next project and worked both with javascript enabled and disabled.
Will run through more test cases on the testbed

<img width="498" alt="image" src="https://github.com/getcord/monorepo/assets/62358728/d20d6aba-5cb1-40ff-8d88-654c673c9a9d">
☝️ 3 working images and one with a url that is broken, so renders the fallback 🙌

Reviewers: AlbertQM, lazybean

Reviewed By: lazybean

Pull Request: https://github.com/getcord/monorepo/pull/8545

monorepo-commit: 33c3743f3420e2fa5d616e497e11a7c9537af360