Message seen observer (#8556)
## Summary
First approach to add a message seen observer to message. This is enough
to make Clack work but there are things left to be done:
- Whenever the user replaces Message, we lose this functionality. We
need to figure out a way of either keeping it transparent for them or
documenting it and exporting the hook.
- The API to mark notifications seen does not exist yet, we need to add
it and call it in `clearNotificationsForMessage` when it's done
- Maybe we should expose a function to be passed by the user that we can
call when the message has been seen?

## Test Plan
Load testbed. Added a console.log in the clear notifications and
markThreadSeen. It was shown when scrolling down.

monorepo-commit: 894bb93204fc687f285715a5c14d2bea5994a377