Add data-cord-message-id and data-cord-thread-id attribute to Message and Thread
This adds `data-cord-message-id` and `data-cord-thread-id` as attributes on the message DOM element for Messages and `data-cord-thread-id` as an attribute on the Thread element.

Of note, I had to pass `ClientMessageData` to `MessageLayout` which did not have this before.  Otherwise, I was able to just add the attribute because the message data was already present.


Test Plan: Verified in testbed

Reviewers: AlbertQM, lazybean, jwatzman

Reviewed By: lazybean, jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/8591

monorepo-commit: b085f2c9a6a0f866c1eaa1c024f06459ebfd9165