Make all location filters take LocationFilterOptions (#8598)
In the JS thread APIs, we generally take either `Location` or
`LocationFilterOptions` for location filters, with the former treated
as exact match.  In the REST APIs and notification APIs, we took only
`Location`, and treated it as partial match.  Unify them by allowing
them all to accept `LocationFilterOptions` and treating a plain
`Location` as exact match.

This is a change of behavior, but looking at our logs, I can't find
any use in the last 7d of any queries that used a location filter in
any of these places.  So this should be safe to change, and it will
make our APIs more consistent.


Test Plan:

npm run test

Added a location filter to a NotificationList component, it showed
notifications from that location.  Added `partialMatch: true` and it
showed partially-matching items.

monorepo-commit: 8f2f9dbb1247ff4ce092675ada1b79cf85142d8d