Revisit props for buttons (#8616)
## Summary

Revisiting props for Button (and SendButton). Instead of just getting
all props from buttons, selecting the most common ones and passing them
down to the base Button component.

This way, we'll be able to document better what they do and control more
the behaviour of the components. If customers need other props, we can
always revisit

## Test plan
Tried in the test bed, and everything behaves the same

monorepo-commit: 128f6f31b38a35a244a66373abb4ba3f54c0d225