Add sortBy location to searchUsers in users api
Adding another option to the user.searchUsers method.
This new param takes in a type (which is only location)
and a location (the location you want to match).

It then sorts the users with those who match the
location at the top of the list. It also put the viewer
at the botton in the location sort. So the viewer
would be inbetween those who have visited, and those
who have not.


Test Plan:
Tested locally in the dev console, was
andrei on the test bed, then changed to flooey.
Used the sortBy for { page: "testbed" } and could
see andrei is the first choice.

Reviewers: flooey, jwatzman

Reviewed By: jwatzman

Pull Request: https://github.com/getcord/monorepo/pull/8501

monorepo-commit: 7b5adab6b300f3c0d24e0b271d1d309e14c894b2