sdk-js React: Make Vanilla-Extract a dependency
I'm in the process of splitting Cord 3.0 and Cord 4.0.
At the end of the 10 PRs stack, I tried building a package (1.37.1-canary.0), but when trying to use that it failed on `import '@vanilla-extract'`, which is a new line in our `dist/index`. Not sure why it's there, could be because I'm `exporting {} from '@vanilla-extract'`.
Discussed with @Flooey, we'll just make VE a dep rather than a devDep.

Test Plan: Release new package (1.37.1-canary.1). The problem is gone.

Reviewers: flooey

Reviewed By: flooey

Pull Request: https://github.com/getcord/monorepo/pull/8658

monorepo-commit: 087ce815d3ee6e08a80f683ef4cdf82ff0381711